# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "rtfd 是自托管的 Sphinx 文档构建与托管服务，本文档描述其 HTTP API（版本号与 assets/VERSION 保持一致）。\n所有接口挂 /rtfd 前缀，响应统一为 {\"success\":bool,\"message\":string,\"data\":any}，\n业务错误同样以该结构返回（HTTP 200），由 customHTTPErrorHandler 统一处理。\n鉴权：管理接口（项目管理、配置查询、导入导出）与项目级接口（详情/更新/删除/导出）采用 HMAC-SHA256 动态签名，\n需随请求携带 X-Rtfd-Ts（Unix 秒）、X-Rtfd-Nonce（随机串）、X-Rtfd-Sign（HMAC-SHA256 签名）三个头；\n签名串 = ts + \"\\n\" + nonce（以 secret 为 HMAC 密钥，与 method/path/body 无关），密钥为 [api] secret 或项目 secret（二者其一即可）。\n可用 `rtfd sign` 生成三个头，并在 Swagger UI 右上角 Authorize 中分别填入 X-Rtfd-Ts / X-Rtfd-Nonce / X-Rtfd-Sign；构建与 webhook 接口使用项目 secret（HMAC/Token）。\n多数接口兼容两种路径风格：/rtfd/{name}/xxx 与 /rtfd/xxx/{name}（文档只列出前者）。",
        "title": "Rtfd API",
        "contact": {},
        "license": {
            "name": "Apache-2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0"
        },
        "version": "2.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/rtfd/import": {
            "post": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "导入 base64 编码的项目配置（由导出接口或 CLI `rtfd project transfer -e` 生成）；\nname 可选，用于改名导入（缺省取配置中的名称）；导入走创建流程，会重新校验并渲染 Caddy 配置。需管理密钥。",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "导入项目配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "base64 编码的项目配置",
                        "name": "export",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "新项目名称，缺省用配置中的名称",
                        "name": "name",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "data.name 为导入后的项目名",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.importData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "配置非法、名称已存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
        "/rtfd/projects": {
            "get": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "返回全部项目名称；verbose=1 时返回完整项目配置数组。需管理密钥。",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "项目列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "为 true/on/1 时返回完整配置（[]lib.OptionsWithResult）",
                        "name": "verbose",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "项目名称数组；verbose 时为完整配置数组",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "密钥未配置或校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "参数与 CLI `rtfd project create` 一致：url 必需，其余字段可选、空值沿用系统默认；创建后自动汇总渲染 Caddy 配置。\n参数可放在表单、query 或 JSON body（键不区分大小写）；需管理密钥。",
                "consumes": [
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "创建项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称（仅小写字母数字与-_.）",
                        "name": "name",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "git 仓库地址（https 或 ssh）",
                        "name": "url",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "分支或tag，缺省用系统 default_branch",
                        "name": "latest",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Python 版本（需在 [py] 分区中定义，如 3.10）",
                        "name": "version",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "语言，多个用英文逗号分隔，如 zh_CN,en",
                        "name": "lang",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "是否单一版本（true/false）",
                        "name": "single",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "文档源目录，默认 docs",
                        "name": "sourcedir",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "依赖文件相对路径（相对文档源目录）",
                        "name": "requirement",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "是否安装依赖（true/false）",
                        "name": "install",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "pip 源索引地址",
                        "name": "index",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Sphinx 构建器：html/dirhtml/singlehtml",
                        "name": "builder",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "项目密钥（构建与 webhook 签名用）",
                        "name": "secret",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "自定义域名（需已解析到本机）",
                        "name": "domain",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "自定义域名证书路径（与 sslkey 成对）",
                        "name": "sslcrt",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "自定义域名证书私钥路径（与 sslcrt 成对）",
                        "name": "sslkey",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "创建成功，data 为项目配置",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/lib.Options"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "参数非法、名称/自定义域名已存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
...
        "/rtfd/{name}/export": {
            "get": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "返回 base64 编码的项目配置，可通过导入接口或 CLI 在其它 rtfd 实例中还原。\nsysmeta=1 时保留系统 meta（如 _installation_id、_webhook_id）。需管理密钥或项目密钥。\n路径别名：GET /rtfd/export/{name}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "导出项目配置",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "为 true/on/1 时保留系统 meta",
                        "name": "sysmeta",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "data.export 为 base64 配置",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/api.exportData"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "项目不存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
        "/rtfd/{name}/info": {
            "get": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "默认返回项目完整配置；key 指定单个字段；build=1 时附带构建集。需管理密钥或项目密钥。\n路径别名：GET /rtfd/info/{name}",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "项目详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "字段名，如 lang、secret、url，返回 data.{key,value}",
                        "name": "key",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "为 true/on/1 时附带构建集（data.buildset）",
                        "name": "build",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "项目配置（key 时为 data.{key,value}）",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.resd"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/lib.OptionsWithResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "default": {
                        "description": "项目不存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
        "/rtfd/{name}/remove": {
            "post": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "删除项目配置与构建结果，并从 Caddy 配置中移除对应站点。需管理密钥或项目密钥。\n路径别名：/rtfd/remove/{name}（POST 与 DELETE 均可）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "删除项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    },
                    "default": {
                        "description": "项目不存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "RtfdNonce": [],
                        "RtfdSign": [],
                        "RtfdTs": []
                    }
                ],
                "description": "删除项目配置与构建结果，并从 Caddy 配置中移除对应站点。需管理密钥或项目密钥。\n路径别名：/rtfd/remove/{name}（POST 与 DELETE 均可）",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "项目管理"
                ],
                "summary": "删除项目",
                "parameters": [
                    {
                        "type": "string",
                        "description": "项目名称",
                        "name": "name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    },
                    "default": {
                        "description": "项目不存在或密钥校验失败",
                        "schema": {
                            "$ref": "#/definitions/api.res"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "api.exportData": {
            "type": "object",
            "properties": {
                "export": {
                    "description": "base64 编码的配置",
                    "type": "string"
                },
                "name": {
                    "description": "项目名称",
                    "type": "string"
                }
            }
        },
        "api.importData": {
            "type": "object",
            "properties": {
                "name": {
                    "description": "导入后的项目名称",
                    "type": "string"
                }
            }
        },
        "api.res": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "api.resd": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "lib.BuilderType": {
            "type": "string",
            "enum": [
                "html",
                "dirhtml",
                "singlehtml"
            ],
            "x-enum-varnames": [
                "HTMLBuilder",
                "DirHTMLBuilder",
                "SingleHTMLBuilder"
            ]
        },
        "lib.Options": {
            "type": "object",
            "properties": {
                "Builder": {
                    "description": "Sphinx构建器，支持html、dirhtml、singlehtml",
                    "allOf": [
                        {
                            "$ref": "#/definitions/lib.BuilderType"
                        }
                    ]
                },
                "CustomDomain": {
                    "description": "自定义域名",
                    "type": "string"
                },
                "DefaultDomain": {
                    "description": "默认域名",
                    "type": "string"
                },
                "GSP": {
                    "description": "git服务提供商（自动填充）",
                    "type": "string"
                },
                "HideGit": {
                    "description": "隐藏git",
                    "type": "boolean"
                },
                "Index": {
                    "description": "pypi仓库",
                    "type": "string"
                },
                "Install": {
                    "description": "是否安装项目",
                    "type": "boolean"
                },
                "IsPublic": {
                    "description": "是否为公开仓库（原type，自动填充）",
                    "type": "boolean"
                },
                "Lang": {
                    "description": "文档语言，以半角逗号分隔多种语言",
                    "type": "string"
                },
                "Latest": {
                    "description": "默认显示的分支",
                    "type": "string"
                },
                "Meta": {
                    "description": "额外配置数据",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "Name": {
                    "description": "项目在数据库中唯一标识名",
                    "type": "string"
                },
                "Requirement": {
                    "description": "依赖包文件，以半角逗号分隔多个文件",
                    "type": "string"
                },
                "SSL": {
                    "description": "自定义域名开启HTTPS（自动填充）",
                    "type": "boolean"
                },
                "SSLPrivate": {
                    "description": "自定义域名的ssl私钥",
                    "type": "string"
                },
                "SSLPublic": {
                    "description": "自定义域名的ssl公钥",
                    "type": "string"
                },
                "Secret": {
                    "description": "webhook secret",
                    "type": "string"
                },
                "ShowNav": {
                    "description": "是否显示导航",
                    "type": "boolean"
                },
                "Single": {
                    "description": "是否单一版本",
                    "type": "boolean"
                },
                "SourceDir": {
                    "description": "文档源文件路径",
                    "type": "string"
                },
                "URL": {
                    "description": "git地址，可以是包含用户名密码的私有仓库",
                    "type": "string"
                },
                "Version": {
                    "description": "使用的python版本，如3、3.10，需在配置文件py分区中定义",
                    "allOf": [
                        {
                            "$ref": "#/definitions/lib.PyVer"
                        }
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "RtfdNonce": {
            "description": "动态签名随机串，须与 X-Rtfd-Ts、X-Rtfd-Sign 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Nonce",
            "in": "header"
        },
        "RtfdSign": {
            "description": "动态签名 HMAC-SHA256(secret, ts+\"\\n\"+nonce)，须与 X-Rtfd-Ts、X-Rtfd-Nonce 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Sign",
            "in": "header"
        },
        "RtfdTs": {
            "description": "动态签名时间戳（Unix 秒），须与 X-Rtfd-Nonce、X-Rtfd-Sign 三者同时携带",
            "type": "apiKey",
            "name": "X-Rtfd-Ts",
            "in": "header"
        }
    }
}